% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_numeric.R
\name{check_numeric}
\alias{check_numeric}
\title{Internal helper to validate numeric scalar inputs}
\usage{
check_numeric(x, name)
}
\arguments{
\item{x}{Value to test}

\item{name}{Name of the value, in order to have a dedicated message}
}
\description{
Internal helper to validate numeric scalar inputs
}
