% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stair_v2.R
\name{create_stair_polygons}
\alias{create_stair_polygons}
\title{Create stair polygon drawing instructions}
\usage{
create_stair_polygons(geometry, riser = TRUE)
}
\arguments{
\item{geometry}{\code{stair_geometry} data frame : columns names are hardcoded hereafter, in order to convert geometry into a polygon list of coordinates.}

\item{riser}{Logical; whether to include riser polygons.}
}
\value{
A list of polygon drawing instructions.
}
\description{
Converts stair geometry into polygon drawing instructions.
}
\keyword{internal}
