% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stair_v2.R
\name{.create_polygon}
\alias{.create_polygon}
\title{Create a polygon drawing instruction}
\usage{
.create_polygon(data, i, x, y)
}
\arguments{
\item{data}{\code{data.frame} containing polygon coordinates.}

\item{i}{Row index.}

\item{x}{Character vector of x coordinate column names.}

\item{y}{Character vector of y coordinate column names.}
}
\value{
A polygon drawing instruction.
}
\description{
Basic converting of 2 vectors and a \code{step} column into a polygon list structure
}
\keyword{internal}
