% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stair_v2.R
\name{.draw_stair_polygon}
\alias{.draw_stair_polygon}
\title{Draw a stair polygon}
\usage{
.draw_stair_polygon(polygon)
}
\arguments{
\item{polygon}{A polygon drawing instruction.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Draw a stair polygon
}
\keyword{internal}
