% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stair.R
\name{.plot_stair_dimensions}
\alias{.plot_stair_dimensions}
\title{Plot stair dimensions}
\usage{
.plot_stair_dimensions(geometry, cex_dimensions = 0.8)
}
\arguments{
\item{geometry}{A stair geometry data frame.}

\item{cex_dimensions}{Character expansion factor for labels.}
}
\description{
Adds cumulative horizontal and vertical dimensions to a stair profile.
}
\details{
This is a helper function used by \code{\link{plot_stair}} when
\code{show_coordinates = TRUE}.
}
\keyword{internal}
