% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giron_blondel.R
\name{giron_blondel}
\alias{giron_blondel}
\title{Calcule le giron optimal selon la formule de Blondel (cible fixe)}
\usage{
giron_blondel(hauteur_marche, cible = 63)
}
\arguments{
\item{hauteur_marche}{Hauteur de marche (cm), typiquement issue de
\code{\link{nombre_marches_optimal}}.}

\item{cible}{Valeur cible de la relation de Blondel \code{2h + g} (cm). Défaut 63.}
}
\value{
Le giron (profondeur de marche) en cm.
}
\description{
Applique la relation \code{2 * hauteur_marche + giron = cible} (cible = 63 cm
par défaut, valeur de confort classique) pour en déduire le giron.
}
\examples{
giron_blondel(16.25)  # 30.5
giron_blondel(20)     # 23

}
