% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landing.R
\name{landing}
\alias{landing}
\title{Create a stair landing object}
\usage{
landing(depth = 0, type = c("none", "start", "end", "intermediate"))
}
\arguments{
\item{depth}{\code{numeric} - Horizontal depth of the landing in mm.}

\item{type}{\code{character} - Landing position:
"none", "start", "end", "intermediate".}
}
\value{
An object of class "landing".
}
\description{
A landing is a horizontal walking surface that does not create a
vertical rise. It can be located at the beginning, end or between
stair flights.
}
