% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_escalier.R
\name{plot.escalier_geometrie}
\alias{plot.escalier_geometrie}
\title{Méthode plot() pour les géométries d'escalier}
\usage{
\method{plot}{escalier_geometrie}(x, ...)
}
\arguments{
\item{x}{Un data.frame de géométrie de classe \code{escalier_geometrie}.}

\item{...}{Arguments passés à \code{\link{plot_escalier}}.}
}
\description{
Permet d'appeler \code{plot(geometrie)} directement sur une géométrie
produite par \code{\link{build_geometry}} (classe
\code{escalier_geometrie}), sans passer par \code{plot.data.frame()} (qui
produirait un pairs-plot illisible). Délègue entièrement à
\code{\link{plot_escalier}}.
}
\examples{
geo <- build_geometry(5, 17.33, rep(28.33, 4))
plot(geo)
plot(geo, cotes = TRUE)

}
