% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene_plot.R
\name{scene_add_treads}
\alias{scene_add_treads}
\title{Add tread segments to a stair scene}
\usage{
scene_add_treads(scene, layout)
}
\arguments{
\item{scene}{A \code{stair_scene}.}

\item{layout}{A rendered geometry.}
}
\value{
The modified scene.
}
\description{
Creates one horizontal segment for each tread.
}
