% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stair_compute.R
\name{stair_compute}
\alias{stair_compute}
\title{Compute complete stair geometry}
\usage{
stair_compute(
  height,
  max_run,
  start = landing(),
  end = landing(),
  first_step_is_floor = FALSE,
  last_step_is_landing = FALSE,
  blondel_target = 630,
  min_going = 150,
  max_going = 350
)
}
\arguments{
\item{height}{\code{numeric} - Total height to overcome in mm.}

\item{max_run}{\code{numeric} - Maximum available horizontal length in mm.}

\item{start}{A \code{landing} object representing the starting landing.}

\item{end}{A \code{landing} object representing the ending landing.}

\item{first_step_is_floor}{\code{logical} - Whether the first level represents a floor.}

\item{last_step_is_landing}{\code{logical} - Whether the arrival level is a landing.}

\item{blondel_target}{\code{numeric} - Target Blondel value in mm.}

\item{min_going}{\code{numeric} - Minimum acceptable tread depth (horizontal depth of one tread) in mm. Default is 150 mm.}

\item{max_going}{\code{numeric} - Maximum acceptable tread depth (horizontal depth of one tread) in mm. Default is 350 mm.}
}
\value{
Object of class "stair_solution".
}
\description{
Calculates a stair geometry fitted to available horizontal space.
}
\examples{
s <- stair_compute(height = 850, max_run = 1030)
}
