% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stair_elements.R
\name{stair_elements}
\alias{stair_elements}
\title{Build stair construction elements}
\usage{
stair_elements(rise, run)
}
\arguments{
\item{rise}{stair_rise object.}

\item{run}{stair_run object.}
}
\value{
A list of stair elements.
}
\description{
Converts stair calculations into a sequence of horizontal and vertical
elements.
}
