% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stair_geometry.R
\name{stair_geometry}
\alias{stair_geometry}
\title{Create stair geometric model}
\usage{
stair_geometry(rise, run, last_step_is_landing = FALSE)
}
\arguments{
\item{rise}{Object of class "stair_rise".}

\item{run}{Object of class "stair_run".}

\item{last_step_is_landing}{\code{logical} - Whether the arrival level is
represented as a landing.}
}
\value{
Object of class "stair_geometry".
}
\description{
Creates the walkable geometry of a stair flight.
}
