% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stair_horizontal_run.R
\name{stair_horizontal_run}
\alias{stair_horizontal_run}
\title{Calculate stair horizontal geometry}
\usage{
stair_horizontal_run(n_rises, going, start = landing(), end = landing())
}
\arguments{
\item{n_rises}{\code{numeric} - Number of vertical rises.}

\item{going}{\code{numeric} - Tread depth in mm.}

\item{start}{A \code{landing} object representing the starting landing.}

\item{end}{A \code{landing} object representing the ending landing.}
}
\value{
An object of class "stair_run".
}
\description{
Calculates tread count, flight length and total horizontal occupation.
}
