% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stair_rise.R
\name{stair_rise}
\alias{stair_rise}
\title{Calculate stair vertical geometry}
\usage{
stair_rise(height, target_rise = 160, min_rise = 160, max_rise = 200)
}
\arguments{
\item{height}{Total height in mm.}

\item{target_rise}{Target rise height.}

\item{min_rise}{Minimum allowed rise.}

\item{max_rise}{Maximum allowed rise.}
}
\value{
Object of class "stair_rise".
}
\description{
Finds the best number of rises for a given height.
}
