% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stair_solve.R
\name{stair_solve}
\alias{stair_solve}
\title{Solve complete stair geometry}
\usage{
stair_solve(
  height,
  max_run,
  start = landing(),
  end = landing(),
  first_step_is_floor = FALSE,
  last_step_is_landing = FALSE
)
}
\arguments{
\item{height}{Total height to overcome in mm.}

\item{max_run}{Maximum available horizontal length in mm.}
}
\value{
Object of class "stair_solution".
}
\description{
Calculates optimal and space-fitted stair solutions.
}
