% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_escalier.R
\name{plot_escalier}
\alias{plot_escalier}
\title{Trace le profil d'un escalier (base R)}
\usage{
plot_escalier(
  geometrie,
  col_standard = "black",
  col_paliere = "red",
  col_sol = "gray50",
  ...
)
}
\arguments{
\item{geometrie}{Un data.frame de géométrie (voir \code{\link{construire_geometrie}}).}

\item{col_standard}{Couleur des girons standards. Défaut "black".}

\item{col_paliere}{Couleur du giron de la marche palière. Défaut "red".}

\item{col_sol}{Couleur des lignes pointillées de sol fini. Défaut "gray50".}

\item{...}{Arguments supplémentaires passés à \code{plot()}.}
}
\description{
Trace chaque contremarche (segment vertical) et chaque giron (segment
horizontal) à partir d'un data.frame de géométrie produit par
\code{\link{construire_geometrie}} (ou \code{calculer_toutes_solutions(...)$par_candidat[[...]]$geometries[[...]]}).
La marche palière (le cas échéant) est mise en évidence dans une couleur
différente. Le sol fini de départ (y=0) et d'arrivée (y=hauteur max) sont
tracés en pointillés pour repère.
}
\details{
Les géométries produites par \code{\link{construire_geometrie}} portent la
classe \code{escalier_geometrie} : appeler \code{plot()} directement dessus
fonctionne aussi (voir \code{\link{plot.escalier_geometrie}}), pas besoin
d'appeler \code{plot_escalier()} par son nom.
}
\examples{
geo <- construire_geometrie(5, 17.33, rep(28.33, 4))
plot_escalier(geo)
plot(geo)  # identique, grâce à la classe escalier_geometrie

# Sur une solution issue de calculer_toutes_solutions() :
sol <- calculer_toutes_solutions(hauteur_a_franchir = 160, distance_max = 1000)
plot_escalier(sol$par_candidat$n10$geometries$avec_palier_max)

}
