% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stair.R
\name{plot_stair}
\alias{plot_stair}
\alias{plot.stair_geometry}
\title{Plot a stair profile}
\usage{
plot_stair(
  geometry,
  col_step = "black",
  col_floor = "gray50",
  show_dimensions = FALSE,
  cex_dimensions = 0.8,
  ...
)

\method{plot}{stair_geometry}(x, ...)
}
\arguments{
\item{geometry}{A data frame of class \code{stair_geometry} returned by
\code{\link{build_geometry}}.}

\item{col_step}{Colour used to draw step segments. Default is
\code{"black"}.}

\item{col_floor}{Colour used for finished floor reference lines.
Default is \code{"gray50"}.}

\item{show_dimensions}{Logical. If \code{TRUE}, cumulative horizontal
and vertical dimensions are displayed. Default is \code{FALSE}.}

\item{cex_dimensions}{Character expansion factor for dimension labels.
Default is \code{0.6}.}

\item{...}{Additional graphical parameters passed to
\code{\link[graphics]{plot}}.}

\item{x}{A \code{stair_geometry} object.}
}
\description{
Draws the profile of a staircase from a geometry data frame produced by
\code{\link{build_geometry}}.
}
\details{
Each step is represented by a vertical rise and a horizontal going.
The lower and upper finished floor levels are displayed as dashed
reference lines.

Objects returned by \code{\link{build_geometry}} inherit from the
\code{stair_geometry} class, allowing direct use of
\code{plot()}.
}
\examples{
geometry <- build_geometry(5, 17.33, rep(28.33, 4))
plot_stair(geometry)

plot(geometry)

plot_stair(geometry, show_dimensions = TRUE)

}
