% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stair.R
\name{plot_stair_dimensions}
\alias{plot_stair_dimensions}
\title{Plot a stair profile with dimensions}
\usage{
plot_stair_dimensions(geometry, ...)
}
\arguments{
\item{geometry}{A stair geometry data frame returned by
\code{\link{build_geometry}}.}

\item{...}{Additional arguments passed to
\code{\link{plot_stair}}.}
}
\description{
Shortcut for \code{plot_stair(..., show_dimensions = TRUE)}.
}
\examples{
geometry <- build_geometry(5, 17.33, rep(28.33, 4))
plot_stair_dimensions(geometry)

}
