% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stair_geometry.R
\name{stair_geometry}
\alias{stair_geometry}
\title{Create stair geometric model}
\usage{
stair_geometry(elements)
}
\arguments{
\item{elements}{Stair element list.}
}
\value{
Object of class "stair_geometry".
}
\description{
Create stair geometric model
}
