% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stair_horizontal_run.R
\name{stair_horizontal_run}
\alias{stair_horizontal_run}
\title{Calculate stair horizontal geometry}
\usage{
stair_horizontal_run(n_rises, going, start = landing(), end = landing())
}
\arguments{
\item{n_rises}{Number of vertical rises.}

\item{going}{Tread depth.}

\item{start}{Start landing object.}

\item{end}{End landing object.}
}
\value{
Object of class "stair_run".
}
\description{
Calculates tread count, flight length and total horizontal occupation.
}
