% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stair_rise.R
\name{stair_rise}
\alias{stair_rise}
\title{Calculate stair vertical geometry}
\usage{
stair_rise(height, target_rise = 160, min_rise = 160, max_rise = 200)
}
\arguments{
\item{height}{\code{numeric} - Total height in mm.}

\item{target_rise}{\code{numeric} - Target rise height in mm.}

\item{min_rise}{\code{numeric} - Minimum allowed rise height in mm.}

\item{max_rise}{\code{numeric} - Maximum allowed rise height in mm.}
}
\value{
Return an object of class "stair_rise".
}
\description{
Determines the number of rises that gives the closest match
to the target rise while respecting minimum and maximum limits.
}
\examples{
stair_rise(3000)

stair_rise(
  height = 1000,
  min_rise = 160,
  max_rise = 250
)
}
