% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stair_solve.R
\name{stair_solve}
\alias{stair_solve}
\title{Solve stair geometry from height and available length}
\usage{
stair_solve(
  height,
  max_run,
  target_rise = 160,
  blondel = 630,
  min_rise = 160,
  max_rise = 200
)
}
\arguments{
\item{height}{Total height in mm.}

\item{max_run}{Maximum available horizontal length in mm.}

\item{target_rise}{Desired rise height.}

\item{blondel}{Target Blondel value.}

\item{min_rise}{Numeric. Minimum allowed rise height in millimetres.
Default is 160 mm.}

\item{max_rise}{Numeric. Maximum allowed rise height in millimetres.
Default is 200 mm.}
}
\value{
Object of class "stair_solution".
}
\description{
Searches for stair solutions using Blondel's formula and available space.
}
\examples{
s <- stair_solve(height = 850,max_run = 1030)

}
